+++ src/components/creategroup/ActivityPeriodSection/ActivityPeriodSection.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useMemo } from 'react';
-import { Section, SectionTitle } from '../../CommonSection.styled';
+import { Section, SectionTitle } from '../../../pages/group/CommonSection.styled';
 import DateWheel from './DateWheel';
 import {
   DatePickerContainer,
