+++ src/components/today-words/MessageList/MessageActionBottomSheet.styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { semanticColors, typography } from '../../../../styles/global/global';
+import { semanticColors, typography } from '../../../styles/global/global';
 
 export const Overlay = styled.div<{ isOpen: boolean }>`
   position: fixed;
