+++ src/pages/group/CreateGroup.tsx
@@ -2,12 +2,12 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import TitleHeader from '../../components/common/TitleHeader';
 import BookSearchBottomSheet from '../../components/common/BookSearchBottomSheet/BookSearchBottomSheet';
-import BookSelectionSection from './components/BookSelectionSection';
-import GenreSelectionSection from './components/GenreSelectionSection';
-import RoomInfoSection from './components/RoomInfoSection';
-import ActivityPeriodSection from './components/ActivityPeriodSection/ActivityPeriodSection';
-import MemberLimitSection from './components/MemberLimitSection';
-import PrivacySettingSection from './components/PrivacySettingSection/PrivacySettingSection';
+import BookSelectionSection from '../../components/creategroup/BookSelectionSection';
+import GenreSelectionSection from '../../components/creategroup/GenreSelectionSection';
+import RoomInfoSection from '../../components/creategroup/RoomInfoSection';
+import ActivityPeriodSection from '../../components/creategroup/ActivityPeriodSection/ActivityPeriodSection';
+import MemberLimitSection from '../../components/creategroup/MemberLimitSection';
+import PrivacySettingSection from '../../components/creategroup//PrivacySettingSection/PrivacySettingSection';
 import leftarrow from '../../assets/common/leftArrow.svg';
 import { Container } from './CreateGroup.styled';
 import { Section } from './CommonSection.styled';
